Isolate legacy WP logic under a dedicated legacy-wp/ directory
Everything a legacy WordPress / legacy PHP setup needs now lives
under packages/playground/wordpress/src/legacy-wp/:

  * legacy-boot.ts — bootLegacyWordPress(), the self-contained
    entry point mirroring bootWordPress's step ordering for PHP
    5.2 + WP 1.0–4.9 on SQLite. Also hosts applyLegacyPhpIniOverrides
    (called from bootRequestHandler).
  * legacy-fixes.ts — WordPress source-file patches
    (patchWordPressSourceFiles, backportWpPreV62MysqlCheck),
    generateDbPhpContent, runPostInstallLegacyFixups, and the
    LEGACY_WP_ERROR_REPORTING_* constants.
  * mysql-shims.ts — PHP shims for mysql_* / mysqli_* APIs that
    old WP expects.

boot.ts stays modern-first: a single
`if (isLegacyPHPVersion(options.phpVersion)) return bootLegacyWordPress(...)`
dispatches at the top of bootWordPress, and the body has zero
legacy conditionals. assertValidDatabaseConnection no longer takes
a phpVersion param — only the modern path reaches it. The
installWordPressSafe + assertValidDatabaseConnectionSafe wrappers
are gone; each boot path owns its own error handling.

Also:

  * Gate backportWpPreV62MysqlCheck explicitly to WP 5.0–6.1 (renamed
    from patchLegacyMysqlCheckForModernWp); skip the placeholder
    db.php write on WP 6.2+ so modern trunk behaviour is preserved.
  * Replace parseInt(phpVersion, 10) < 7 checks with
    isLegacyPHPVersion() so non-numeric values like 'latest' never
    slip into the legacy branch.
