+++ packages/playground/wordpress/src/boot.ts
@@ -9,6 +9,7 @@ import type {
 	Remote,
 } from '@php-wasm/universal';
 import {
+	isLegacyPHPVersion,
 	PHP,
 	PHPRequestHandler,
 	sandboxedSpawnHandlerFactory,
@@ -26,78 +27,12 @@ import {
 import { basename, dirname, joinPaths } from '@php-wasm/util';
 import { logger } from '@php-wasm/logger';
 import { ensureWpConfig } from './wp-config';
+import { assertDatabasePrerequisites } from './database-prerequisites';
 import {
-	generateDbPhpContent,
-	LEGACY_WP_ERROR_REPORTING_PHP_EXPR,
-	LEGACY_WP_ERROR_REPORTING_VALUE,
-	patchWordPressSourceFiles,
-	runPostInstallLegacyFixups,
-} from './legacy-wp-fixes';
-
-/**
- * Network I/O functions that must be disabled on legacy PHP builds
- * (< 7) to avoid "null function or function signature mismatch"
- * WASM crashes when WordPress calls fsockopen or cURL during cron,
- * update checks, dashboard RSS widgets, etc.
- */
-const LEGACY_PHP_DISABLED_NETWORK_FUNCTIONS = [
-	'fsockopen',
-	'pfsockopen',
-	'curl_init',
-	'curl_exec',
-	'curl_multi_exec',
-] as const;
-
-/**
- * Minimal wp-content/db.php drop-in for modern WordPress. Its job
- * is simply to exist, so WP's file_exists(WP_CONTENT_DIR . '/db.php')
- * escape hatches (in wp_check_php_mysql_versions() and install.php
- * step=2's mysql version check) both fall through. The real SQLite
- * wiring is handled by the preloaded lazy $wpdb loader, not by this
- * file — so the content is deliberately an empty, marked placeholder
- * that WordPress require()s as a no-op.
- */
-const MODERN_PLAYGROUND_DB_PHP = `<?php
-// @playground-managed — Playground-generated db.php placeholder.
-//
-// The SQLite $wpdb is set up by the preloaded lazy loader via
-// auto_prepend_file. This file only needs to exist to satisfy
-// WordPress' own file_exists escape hatches in its MySQL checks.
-`;
-
-/**
- * Backports WP 6.2's wp_check_php_mysql_versions() MySQL check to
- * older WordPress releases on modern PHP.
- *
- * WP 5.0–6.1 hard-check `extension_loaded('mysqli')`, which a userland
- * stub cannot satisfy — so when Playground boots the SQLite-backed
- * preload, WordPress still dies with "Requirements Not Met — missing
- * MySQL extension" on these versions. WP 6.2+ switched to
- * `function_exists('mysqli_connect')`, which the preload's stub
- * already provides.
- *
- * Rewrite the single `extension_loaded('mysqli')` call in
- * wp-includes/load.php to `function_exists('mysqli_connect')`, in
- * place. The patch is a no-op on WP 6.2+ (pattern won't match) and
- * on any WordPress release that already uses function_exists.
- */
-async function patchLegacyMysqlCheckForModernWp(
-	php: PHP,
-	documentRoot: string
-): Promise<void> {
-	const loadPhp = joinPaths(documentRoot, 'wp-includes/load.php');
-	if (!php.fileExists(loadPhp)) {
-		return;
-	}
-	const content = php.readFileAsText(loadPhp);
-	const patched = content.replace(
-		"extension_loaded( 'mysqli' )",
-		"function_exists( 'mysqli_connect' )"
-	);
-	if (patched !== content) {
-		await php.writeFile(loadPhp, patched);
-	}
-}
+	applyLegacyPhpIniOverrides,
+	bootLegacyWordPress,
+} from './legacy-wp/legacy-boot';
+import { backportWpPreV62MysqlCheck } from './legacy-wp/legacy-fixes';
 
 export type PhpIniOptions = Record<string, string>;
 export type Hook = (php: PHP) => void | Promise<void>;
@@ -288,6 +223,10 @@ export async function bootWordPress(
 	requestHandler: PHPRequestHandler,
 	options: BootWordPressOptions
 ) {
+	if (isLegacyPHPVersion(options.phpVersion)) {
+		return bootLegacyWordPress(requestHandler, options);
+	}
+
 	const php = await requestHandler.getPrimaryPhp();
 	if (options.hooks?.beforeWordPressFiles) {
 		await options.hooks.beforeWordPressFiles(php);
@@ -316,41 +255,7 @@ export async function bootWordPress(
 	 * them. This is needed because some WordPress backups and exports may not
 	 * include definitions for some of the necessary constants.
 	 */
-	const phpMajor = Number.isFinite(parseInt(options.phpVersion ?? '', 10))
-		? parseInt(options.phpVersion!, 10)
-		: 8;
-	if (phpMajor >= 7) {
-		await ensureWpConfig(php, requestHandler.documentRoot);
-	} else {
-		// For legacy PHP, skip ensureWpConfig since the pre-built
-		// WordPress already has a valid wp-config-sample.php and
-		// php.run() with the large transformer code hangs.
-		// Just copy wp-config-sample.php to wp-config.php if needed.
-		const wpConfigPath = joinPaths(
-			requestHandler.documentRoot,
-			'wp-config.php'
-		);
-		if (
-			!php.fileExists(wpConfigPath) &&
-			php.fileExists(
-				joinPaths(requestHandler.documentRoot, 'wp-config-sample.php')
-			)
-		) {
-			await php.writeFile(
-				wpConfigPath,
-				await php.readFileAsBuffer(
-					joinPaths(
-						requestHandler.documentRoot,
-						'wp-config-sample.php'
-					)
-				)
-			);
-		}
-	}
-	if (phpMajor < 7) {
-		await patchWordPressSourceFiles(php, requestHandler.documentRoot);
-	}
-
+	await ensureWpConfig(php, requestHandler.documentRoot);
 	// Run "before database" hooks to mount/copy more files in
 	if (options.hooks?.beforeDatabaseSetup) {
 		await options.hooks.beforeDatabaseSetup(php);
@@ -366,51 +271,7 @@ export async function bootWordPress(
 			await options.sqliteIntegrationPluginZip,
 			{ phpVersion: options.phpVersion }
 		);
-
-		// Write wp-content/db.php. Two distinct WordPress code paths
-		// rely on the file_exists(WP_CONTENT_DIR . '/db.php') escape
-		// hatch and cannot be fixed by a userland function stub:
-		//
-		//   * wp-includes/load.php :: wp_check_php_mysql_versions()
-		//     uses extension_loaded('mysqli') on WP 5.0–6.1.
-		//   * wp-admin/install.php step=2 has its own mysql version
-		//     check that compares $wpdb->db_version() against
-		//     $required_mysql_version and falls through on db.php
-		//     existence. WP_SQLite_DB's reported version does not
-		//     satisfy WP's mysql minimum, so install dies silently
-		//     without this file.
-		//
-		// Legacy PHP needs a full-content db.php because old
-		// WordPress actually uses the MySQL function stubs from it.
-		// Modern PHP only needs the file to exist — the preload's
-		// lazy $wpdb loader still owns the real connection, as long
-		// as the preload's guard recognises our @playground-managed
-		// marker and does not self-skip on our own file.
-		const wpContentDir = joinPaths(
-			requestHandler.documentRoot,
-			'wp-content'
-		);
-		const dbPhpPath = joinPaths(wpContentDir, 'db.php');
-		if (php.isDir(wpContentDir) && !php.fileExists(dbPhpPath)) {
-			await php.writeFile(
-				dbPhpPath,
-				phpMajor < 7 ? generateDbPhpContent() : MODERN_PLAYGROUND_DB_PHP
-			);
-		}
-
-		// WordPress 5.0–6.1's `wp_check_php_mysql_versions()` runs
-		// before `wp_initial_constants()` defines WP_CONTENT_DIR, so
-		// the file_exists escape hatch above alone cannot rescue the
-		// early check — the path collapses to 'WP_CONTENT_DIR/db.php'
-		// via an undefined constant. Patch wp-includes/load.php to
-		// use function_exists('mysqli_connect') instead, matching the
-		// fix WordPress itself shipped in 6.2. Also a no-op on 6.2+.
-		if (phpMajor >= 7) {
-			await patchLegacyMysqlCheckForModernWp(
-				php,
-				requestHandler.documentRoot
-			);
-		}
+		await backportWpPreV62MysqlCheck(php, requestHandler.documentRoot);
 	}
 
 	const installationMode =
@@ -428,18 +289,20 @@ export async function bootWordPress(
 			hasCustomDatabasePath,
 		});
 		// Install WordPress if it's not installed.
-		await installWordPressSafe(
-			php,
-			phpMajor,
-			hasCustomDatabasePath,
-			requestHandler,
-			options.phpVersion
-		);
+		try {
+			await installWordPress(php);
+		} catch (error) {
+			// If installation failed, check if it's a database issue
+			// to provide a more specific error message (but skip if user provided custom DB path)
+			if (!hasCustomDatabasePath) {
+				await assertValidDatabaseConnection(requestHandler);
+			}
+			// If we get here, the database is valid but installation failed for another reason
+			throw error;
+		}
+		// Validate the database connection after installation (skip if user provided custom DB path)
 		if (!hasCustomDatabasePath) {
-			await assertValidDatabaseConnectionSafe(
-				requestHandler,
-				options.phpVersion
-			);
+			await assertValidDatabaseConnection(requestHandler);
 		}
 	} else if ('install-from-existing-files-if-needed' === installationMode) {
 		// Check database prerequisites before attempting installation
@@ -447,142 +310,27 @@ export async function bootWordPress(
 			usesSqlite,
 			hasCustomDatabasePath,
 		});
-		// For legacy PHP (< 7), skip isWordPressInstalled check because
-		// it crashes the WASM runtime on old WordPress (< 3.0) where the
-		// SQLite driver initialization chain isn't fully compatible.
-		const isInstalled =
-			phpMajor >= 7 ? await isWordPressInstalled(php) : false;
-		if (!isInstalled) {
-			await installWordPressSafe(
-				php,
-				phpMajor,
-				hasCustomDatabasePath,
-				requestHandler,
-				options.phpVersion
-			);
+		if (!(await isWordPressInstalled(php))) {
+			// Install WordPress if it's not installed.
+			try {
+				await installWordPress(php);
+			} catch (error) {
+				// If installation failed, check if it's a database issue
+				// to provide a more specific error message (but skip if user provided custom DB path)
+				if (!hasCustomDatabasePath) {
+					await assertValidDatabaseConnection(requestHandler);
+				}
+				// If we get here, the database is valid but installation failed for another reason
+				throw error;
+			}
 		}
-		// Validate the database connection after installation
+		// Validate the database connection after installation (skip if user provided custom DB path)
 		if (!hasCustomDatabasePath) {
-			await assertValidDatabaseConnectionSafe(
-				requestHandler,
-				options.phpVersion
-			);
+			await assertValidDatabaseConnection(requestHandler);
 		}
 	}
 
 	return requestHandler;
-}
-
-/**
- * Wrapper around installWordPress that handles errors gracefully
- * for legacy PHP versions where installation errors may be non-fatal.
- */
-async function installWordPressSafe(
-	php: PHP,
-	phpMajor: number,
-	hasCustomDatabasePath: boolean,
-	requestHandler: PHPRequestHandler,
-	phpVersion?: string
-): Promise<void> {
-	try {
-		await installWordPress(php, phpMajor);
-	} catch (error) {
-		if (!hasCustomDatabasePath) {
-			await assertValidDatabaseConnectionSafe(requestHandler, phpVersion);
-		}
-		if (phpMajor >= 7) {
-			throw error;
-		}
-		logger.warn('Legacy PHP WordPress installation error:', error);
-	}
-	// Run legacy fixups whether the installer succeeded or threw. On
-	// WP 1.x the installer routinely fails halfway through and we rely
-	// on the fixups (stage 2 in particular) to finish building the
-	// schema. On newer legacy WP where the installer succeeded, the
-	// fixups short-circuit cheaply: stage 1 exits before loading WP if
-	// wp_users doesn't exist yet, and stage 2 is gated to WP < 3.5, so
-	// the only work done on the happy path is a pair of UPDATE queries
-	// against wp_options (siteurl/home) plus an admin-password reset.
-	if (phpMajor < 7) {
-		await runPostInstallLegacyFixups(php, requestHandler.absoluteUrl);
-	}
-}
-
-/**
- * Checks if database prerequisites are in place before attempting WordPress installation.
- * This performs lightweight checks that don't require WordPress to be installed.
- */
-async function assertDatabasePrerequisites(
-	requestHandler: PHPRequestHandler,
-	{
-		usesSqlite,
-		hasCustomDatabasePath,
-	}: {
-		usesSqlite: boolean;
-		hasCustomDatabasePath: boolean;
-	}
-) {
-	const php = await requestHandler.getPrimaryPhp();
-
-	// If SQLite integration is preloaded via core, we're good
-	if (php.isFile('/internal/shared/preload/0-sqlite.php')) {
-		return;
-	}
-
-	// Check if a SQLite integration plugin directory exists (even if not provided via zip)
-	// This handles cases where the directory is mounted via hooks
-	const sqlitePluginPath = joinPaths(
-		requestHandler.documentRoot,
-		'wp-content/mu-plugins/sqlite-database-integration'
-	);
-
-	if (php.isDir(sqlitePluginPath)) {
-		// The directory exists, we'll validate it after WordPress is installed
-		return;
-	}
-
-	// Check if we provided a SQLite integration zip
-	if (usesSqlite) {
-		// We provided a zip, so SQLite will be set up during boot
-		return;
-	}
-
-	// If we have a custom database path (dataSqlPath option was provided),
-	// assume it's configured - the actual connection will be validated after installation
-	if (hasCustomDatabasePath) {
-		return;
-	}
-
-	// Check if wp-config.php has real MySQL credentials
-	if (hasValidMySQLCredentials(php)) {
-		return;
-	}
-
-	// No SQLite integration and no MySQL credentials found
-	// Throw early to avoid attempting installation with no database
-	throw new Error('Error connecting to the MySQL database.');
-}
-
-/**
- * For legacy PHP (< 7), skip the database connection check entirely.
- *
- * Calling isDatabaseConnectionValid() loads wp-load.php. On some old
- * WordPress versions (2.5–2.7) this triggers a WASM "null function or
- * function signature mismatch" crash that corrupts the PHP instance and
- * prevents the front page from loading. The check is non-fatal for
- * legacy PHP anyway — runPostInstallLegacyFixups() handles any setup
- * that's needed. Skipping gives the same observable result (no error
- * thrown) without the risk of state corruption.
- */
-async function assertValidDatabaseConnectionSafe(
-	requestHandler: PHPRequestHandler,
-	phpVersion?: string
-) {
-	const phpMajor = parseInt(phpVersion ?? '8', 10);
-	if (phpMajor < 7) {
-		return;
-	}
-	await assertValidDatabaseConnection(requestHandler);
 }
 
 async function assertValidDatabaseConnection(
@@ -637,41 +385,12 @@ export async function bootRequestHandler(options: BootRequestHandlerOptions) {
 			setPhpIniEntries(php, options.phpIniEntries);
 		}
 
-		// Disable network I/O for legacy PHP (< 7) to prevent WASM
-		// crashes. Old WordPress (2.5–3.6) calls fsockopen/cURL during
-		// cron, update checks, and dashboard RSS widgets. The
-		// underlying socket/cURL operations trigger "null function or
-		// function signature mismatch" WASM errors; disabling them
-		// makes the calls fail safely (return false) instead of
-		// crashing.
-		//
-		// setPhpIniEntries overwrites keys, so we merge with whatever
-		// the caller already passed in `options.phpIniEntries` —
-		// otherwise a networking-disabled list from the web worker
-		// would be silently replaced by this legacy-only list.
-		if (parseInt(options.phpVersion ?? '8', 10) < 7) {
-			const legacyDisabled = [...LEGACY_PHP_DISABLED_NETWORK_FUNCTIONS];
-			const callerDisabled = (
-				options.phpIniEntries?.['disable_functions'] ?? ''
-			)
-				.split(',')
-				.map((s) => s.trim())
-				.filter((s) => s);
-			const mergedDisabled = Array.from(
-				new Set([...callerDisabled, ...legacyDisabled])
-			).join(',');
-			setPhpIniEntries(php, {
-				disable_functions: mergedDisabled,
-				allow_url_fopen: '0',
-			});
-		}
+		applyLegacyPhpIniOverrides(php, {
+			phpVersion: options.phpVersion,
+			phpIniEntries: options.phpIniEntries,
+		});
 
-		// Use the new AST-based SQLite driver for all supported PHP
-		// versions. The PHP 5.2 build of the driver is the
-		// `v2.2.22-php52` variant bundled alongside trunk; it has
-		// closures hoisted to named functions and a few polyfills so
-		// it runs unmodified on PHP 5.2. See the
-		// `sqlite-database-integration-v2.2.22-php52.zip` asset.
+		// Use the new AST-based SQLite driver.
 		// TODO: Remove this once the new driver is the default; when this is closed:
 		//         https://github.com/WordPress/sqlite-database-integration/issues/195
 		php.defineConstant('WP_SQLITE_AST_DRIVER', true);
@@ -820,69 +539,13 @@ export async function isWordPressInstalled(php: PHP) {
  * Without them, the installer may take 60 seconds,
  * 300 seconds, or even more to complete.
  */
-async function installWordPress(php: PHP, phpMajor = 8) {
-	// WP 1.0–3.0 on legacy PHP: skip the install.php HTTP request
-	// entirely. These old installers trigger various unreachable WASM
-	// traps (mail(), mysql_get_server_info(), etc.) that the PHP 5.2
-	// binary can't handle. The runPostInstallLegacyFixups() PDO
-	// fallback creates all tables, users, options, and content
-	// without running any crashable PHP.
-	if (phpMajor < 7) {
-		const versionPhp = joinPaths(
-			php.documentRoot,
-			'wp-includes/version.php'
-		);
-		if (php.fileExists(versionPhp)) {
-			const content = php.readFileAsText(versionPhp);
-			const match = content.match(/\$wp_version\s*=\s*['"]([^'"]+)['"]/);
-			if (match) {
-				const wpVersion = match[1];
-				// WP 1.0–3.0 installers trigger unreachable WASM
-				// traps from mail(), network calls,
-				// mysql_get_server_info(), etc. WP 3.1+ works
-				// with targeted function patches.
-				//
-				// WP 1.0-1.2: the post-install PDO fallback
-				//   creates the very simple schema entirely.
-				// WP 1.5-3.0: needs dbDelta() for proper table
-				//   schemas but skip the rest of the installer.
-				if (parseFloat(wpVersion) < 2.1) {
-					return;
-				}
-				if (parseFloat(wpVersion) <= 3.0) {
-					await runDbDeltaOnly(php);
-					return;
-				}
-			}
-		}
-	}
-
-	const iniOverrides: Record<string, string> = {
-		// Disable network I/O functions during installation.
-		// For legacy PHP (< 7), this must include all the functions
-		// already disabled in bootRequestHandler — setPhpIniEntries
-		// replaces the entire value, so listing only 'fsockopen'
-		// would re-enable curl_init/curl_exec and cause WASM crashes
-		// when the installer makes outbound HTTP requests.
-		disable_functions:
-			phpMajor < 7
-				? [...LEGACY_PHP_DISABLED_NETWORK_FUNCTIONS, 'mail'].join(',')
-				: 'fsockopen',
-		allow_url_fopen: '0',
-	};
-	if (phpMajor < 7) {
-		// Suppress E_DEPRECATED (8192) and E_STRICT (2048) at
-		// the ini level. Old WordPress class declarations trigger
-		// E_STRICT warnings during compilation (e.g. Walker_Page)
-		// which PHP may report using the ini error_reporting value
-		// rather than the runtime error_reporting() call.
-		iniOverrides['error_reporting'] = String(
-			LEGACY_WP_ERROR_REPORTING_VALUE
-		);
-	}
+async function installWordPress(php: PHP) {
 	const response = await withPHPIniValues(
 		php,
-		iniOverrides,
+		{
+			disable_functions: 'fsockopen',
+			allow_url_fopen: '0',
+		},
 		async () =>
 			await php.request({
 				url: '/wp-admin/install.php?step=2',
@@ -902,27 +565,7 @@ async function installWordPress(php: PHP, phpMajor = 8) {
 			})
 	);
 
-	if (phpMajor < 7) {
-		// Legacy PHP (< 7): skip isWordPressInstalled() entirely — it
-		// can trigger a WASM trap (not a PHP exception) on old WordPress
-		// (< 3.0), which corrupts the runtime beyond recovery. Use the
-		// installer response text as a heuristic instead.
-		const installSucceeded =
-			response.text?.includes('Success') ||
-			response.text?.includes('successful') ||
-			response.text?.includes('Finished') ||
-			response.text?.includes('Already Installed') ||
-			response.text?.includes('already have WordPress installed') ||
-			false;
-		if (!installSucceeded) {
-			throw new Error(
-				`Failed to install WordPress – installer responded with "${response.text?.substring(
-					0,
-					100
-				)}"`
-			);
-		}
-	} else if (!(await isWordPressInstalled(php))) {
+	if (!(await isWordPressInstalled(php))) {
 		throw new Error(
 			`Failed to install WordPress – installer responded with "${response.text?.substring(
 				0,
@@ -931,130 +574,35 @@ async function installWordPress(php: PHP, phpMajor = 8) {
 		);
 	}
 
-	if (phpMajor < 7) {
-		// Legacy PHP: set permalink_structure via PDO instead of
-		// update_option(). On WP < 4.8.3, wpdb::prepare() passes
-		// the value through vsprintf() without escaping '%'
-		// characters first (the placeholder_escape mechanism was
-		// added in 4.8.3). The '%y', '%m', '%d', '%p' sequences
-		// in the permalink pattern are interpreted as sprintf
-		// format specifiers, mangling the stored value.
-		// Using PDO bypasses wpdb entirely.
-		try {
-			const result = await php.run({
-				code: `<?php
-					$db_dir = getenv('DOCUMENT_ROOT') . '/wp-content/database/';
-					$db_path = $db_dir . '.ht.sqlite';
-					if (!file_exists($db_path)) { echo '0'; exit; }
-					$pdo = new PDO('sqlite:' . $db_path);
-					$pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-					$nice_permalinks = '/%year%/%monthnum%/%day%/%postname%/';
-					$stmt = $pdo->prepare(
-						"UPDATE wp_options SET option_value = :val WHERE option_name = 'permalink_structure'"
-					);
-					$stmt->execute(array(':val' => $nice_permalinks));
-					if ($stmt->rowCount() === 0) {
-						$stmt = $pdo->prepare(
-							"INSERT INTO wp_options (option_name, option_value, autoload) VALUES ('permalink_structure', :val, 'yes')"
-						);
-						$stmt->execute(array(':val' => $nice_permalinks));
-					}
-					$check = $pdo->query(
-						"SELECT option_value FROM wp_options WHERE option_name = 'permalink_structure'"
-					)->fetchColumn();
-					echo $check === $nice_permalinks ? '1' : '0';
-				`,
-				env: { DOCUMENT_ROOT: php.documentRoot },
-			});
-			if (result.text !== '1') {
-				logger.warn(
-					'Failed to default to pretty permalinks after WP install.'
-				);
+	const defaultedToPrettyPermalinks = await php.run({
+		code: `<?php
+			ob_start();
+			$wp_load = getenv('DOCUMENT_ROOT') . '/wp-load.php';
+			if (!file_exists($wp_load)) {
+				echo '0';
+				exit;
 			}
-		} catch {
-			logger.warn(
-				'Failed to set pretty permalinks after WP install (non-fatal).'
+			require $wp_load;
+			$nice_permalinks = '/%year%/%monthnum%/%day%/%postname%/';
+			$option_result = update_option(
+				'permalink_structure',
+				$nice_permalinks
 			);
-		}
-	} else {
-		const defaultedToPrettyPermalinks = await php.run({
-			code: `<?php
-				ob_start();
-				$wp_load = getenv('DOCUMENT_ROOT') . '/wp-load.php';
-				if (!file_exists($wp_load)) {
-					echo '0';
-					exit;
-				}
-				require $wp_load;
-				$nice_permalinks = '/%year%/%monthnum%/%day%/%postname%/';
-				$option_result = update_option(
-					'permalink_structure',
-					$nice_permalinks
-				);
-				ob_clean();
-				if ( get_option( 'permalink_structure' ) === $nice_permalinks ) {
-					echo '1';
-				} else {
-					echo '0';
-				}
-				ob_end_flush();
-			`,
-			env: {
-				DOCUMENT_ROOT: php.documentRoot,
-			},
-		});
+			ob_clean();
+			if ( get_option( 'permalink_structure' ) === $nice_permalinks ) {
+				echo '1';
+			} else {
+				echo '0';
+			}
+			ob_end_flush();
+		`,
+		env: {
+			DOCUMENT_ROOT: php.documentRoot,
+		},
+	});
 
-		if (defaultedToPrettyPermalinks.text !== '1') {
-			logger.warn(
-				'Failed to default to pretty permalinks after WP install.'
-			);
-		}
-	}
-}
-
-/**
- * Runs dbDelta() and populate_options/populate_roles without the
- * full wp_install() function. Used for WP 2.3–3.0 where the
- * installer crashes but we still need the table schemas.
- */
-async function runDbDeltaOnly(php: PHP): Promise<void> {
-	try {
-		await php.run({
-			code: `<?php
-				define('WP_INSTALLING', true);
-				error_reporting(${LEGACY_WP_ERROR_REPORTING_PHP_EXPR});
-				ini_set('display_errors', '0');
-				ob_start();
-				require getenv('DOCUMENT_ROOT') . '/wp-load.php';
-				ob_clean();
-				// Load upgrade functions for dbDelta
-				if (file_exists(ABSPATH . 'wp-admin/includes/upgrade.php')) {
-					require_once ABSPATH . 'wp-admin/includes/upgrade.php';
-				} elseif (file_exists(ABSPATH . 'wp-admin/upgrade-functions.php')) {
-					require_once ABSPATH . 'wp-admin/upgrade-functions.php';
-				}
-				// Create tables via dbDelta — the critical step that
-				// creates the proper schema for the WP version.
-				if (function_exists('make_db_current_silent')) {
-					make_db_current_silent();
-				}
-				// populate_options/populate_roles on WP 2.3+ only.
-				// WP 2.1-2.2 crash in these functions (WASM traps
-				// from mail/network calls that bypass PHP try/catch).
-				// The PDO fallback seeds essential options/roles.
-				global $wp_version;
-				// populate_options sets db_version and other essential
-				// options. populate_roles creates the roles/capabilities.
-				// On PHP 5.2 WP 2.1-2.2 these crash with WASM traps.
-				// Run them for any WP version that defines them.
-				if (function_exists('populate_options')) populate_options();
-				if (function_exists('populate_roles')) populate_roles();
-				echo 'OK';
-			`,
-			env: { DOCUMENT_ROOT: php.documentRoot },
-		});
-	} catch (error) {
-		logger.warn('runDbDeltaOnly failed (non-fatal):', error);
+	if (defaultedToPrettyPermalinks.text !== '1') {
+		logger.warn('Failed to default to pretty permalinks after WP install.');
 	}
 }
 
@@ -1068,24 +616,6 @@ export function getFileNotFoundActionForWordPress(
 		type: 'internal-redirect',
 		uri: '/index.php',
 	};
-}
-
-function hasValidMySQLCredentials(php: PHP) {
-	const wpConfigPath = joinPaths(php.documentRoot, 'wp-config.php');
-	if (!php.isFile(wpConfigPath)) return false;
-
-	const wpConfig = php.readFileAsText(wpConfigPath);
-
-	const dbName = wpConfig.match(
-		/define\s*\(\s*['"]DB_NAME['"]\s*,\s*['"]([^'"]*)['"]/
-	);
-	const dbUser = wpConfig.match(
-		/define\s*\(\s*['"]DB_USER['"]\s*,\s*['"]([^'"]*)['"]/
-	);
-
-	if (!dbName || !dbUser) return false;
-
-	return dbName[1] !== 'database_name_here' && dbUser[1] !== 'username_here';
 }
 
 async function isDatabaseConnectionValid(php: PHP) {
